(PUP-3526) Add error message about indirection not matching prefix
There are now two ways to have a request get to v3/indirected_routes:
`/puppet/v3` and `/ca/v1`. Thus, a request to `/puppet/v3/certificates` would
be routed to indirected_routes and could be handled successfully, although
issues would probably come up around authorizing the request. Such a request
is incorrect - it has the wrong url prefix - and we should say so.